requirements.txt
kagglehub
numpy
opencv-python
pillow
streamlit
torch
torchvision